refactor(frontend): share POSIX shell quoting
